Restructure usb file handling
Require the caller of Stim300Driver to perform the opening and closing
of the serial driver and handling the errors as they please.
