+++ src/serial_unix.h
@@ -7,23 +7,36 @@
 #include <fcntl.h>
 #include <unistd.h>
 
-#include <string.h>
+#include <cstring>
 #include <stdexcept>
+
+enum class BAUDRATE : uint32_t
+{
+  BAUD_4800 = 4800,
+  BAUD_9600 = 9600,
+  BAUD_19200 = 19200,
+  BAUD_38400 = 38400,
+  BAUD_57600 = 57600,
+  BAUD_115200 = 115200,
+  BAUD_921600 = 921600
+};
 
 class SerialUnix : public SerialDriver
 {
 public:
   explicit SerialUnix(const std::string& serial_port_name);
 
-  ~SerialUnix() override;
+  ~SerialUnix();
 
-  void open(BAUDRATE baudrate) override;
+  void open(BAUDRATE baudrate);
 
-  void close() override;
+  void close();
 
   void writeByte(uint8_t byte) override;
 
   bool readByte(uint8_t& byte) override;
+
+  bool flush() override;
 
 private:
   int file_handle_;
